refactor js entrypoint over to ts
